# backend/docs/openapi.yaml
components:
...
  securitySchemes:
    apiKeyAuth:
      description: API key for device authentication. Provide the API key as a Bearer
        token.
      in: header
      name: Authorization
      type: apiKey
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact:
    name: MOTO Support
  description: API for the MOTO school management system
  title: MOTO API
  version: 1.0.0
openapi: 3.0.3
paths:
...
  /api/active/groups/room/{roomId}:
    get:
      description: Generated from routes
      parameters:
      - description: roomId parameter
        in: path
        name: roomId
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/active/groups/room/{roomId}
      tags:
      - Active
...
  /api/iot/rooms/available:
    get:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/iot/rooms/available
      tags:
      - Iot
...
  /api/rooms:
    get:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/rooms
      tags:
      - Rooms
    post:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: POST /api/rooms
      tags:
      - Rooms
  /api/rooms/{id}:
    delete:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: DELETE /api/rooms/{id}
      tags:
      - Rooms
    get:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/rooms/{id}
      tags:
      - Rooms
    put:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: PUT /api/rooms/{id}
      tags:
      - Rooms
  /api/rooms/{id}/history:
    get:
      description: Generated from routes
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/rooms/{id}/history
      tags:
      - Rooms
  /api/rooms/available:
    get:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/rooms/available
      tags:
      - Rooms
  /api/rooms/buildings:
    get:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/rooms/buildings
      tags:
      - Rooms
  /api/rooms/by-category:
    get:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/rooms/by-category
      tags:
      - Rooms
  /api/rooms/categories:
    get:
      description: Generated from routes
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
      summary: GET /api/rooms/categories
      tags:
      - Rooms
...
servers:
- description: API Base URL
  url: /api